Appel de la fonction PDO
