+++ allobobo_bdd.php
@@ -14,4 +14,6 @@ catch(PDOException $e)
 	exit('Erreur : ' . $e->getMessage());
 }
 
+
 ?>
+
