+++ config.php
@@ -1,0 +1,9 @@
+<?php 
+
+//chemin accès
+define('ALLOBOBO_ROOT', __DIR__ . '/');
+define('ALLOBOBO_DIRNAME',dirname(ALLOBOBO_ROOT));
+define('ALLOBOBO_DB', ALLOBOBO_ROOT . 'PDOconnect.php');
+
+
+
