+++ config/config.php
@@ -2,3 +2,4 @@
 
 // a renomer en .env pour empecher la futur visibilite par git
 define('ADMIN_PASSWORD', 'admin');
+
