+++ dao/DataSqlObject.php
@@ -1,0 +1,4 @@
+<?php 
+/**
+ * ORM & entity manager fait maison
+ */
